refactor: extract helper to fetch json data
