+++ entities/product.ts
